# docs/openapi/v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parity v1 \u2014 Deterministic Financial Normalisation API",
    "version": "1.0.0",
    "description": "Integer-only money. Deterministic hashing. Immutable snapshots."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/v1/deals": {
      "post": {
        "operationId": "createDeal",
        "summary": "Create a new deal",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "currency"
                ],
                "properties": {
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3,
                    "description": "ISO 4217 currency code"
                  },
                  "name": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "created_by": {
                    "format": "uuid",
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deal created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listDeals",
        "summary": "List deals for a user",
        "parameters": [
          {
            "name": "created_by",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deals list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealsListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/v1/deals/{deal_id}": {
      "get": {
        "operationId": "getDeal",
        "summary": "Get deal with analysis runs and snapshots",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deal detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/deals/{deal_id}/documents": {
      "post": {
        "operationId": "uploadDocument",
        "summary": "Upload a financial document for ingestion",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "created_by": {
                    "format": "uuid",
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingestion result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/CurrencyMismatch"
          },
          "400": {
            "$ref": "#/components/responses/InvalidSchema"
          }
        }
      },
      "get": {
        "operationId": "listDocuments",
        "summary": "List documents for a deal",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documents list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsListResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/documents/{document_id}/status": {
      "get": {
        "operationId": "getDocumentStatus",
        "summary": "Get document processing status",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/documents/{document_id}/transactions": {
      "get": {
        "operationId": "getDocumentTransactions",
        "summary": "Get raw transactions for a document",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/deals/{deal_id}/overrides": {
      "post": {
        "operationId": "addOverride",
        "summary": "Add an override (insert-only, append-only log)",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "entity_id",
                  "weight",
                  "new_value"
                ],
                "properties": {
                  "entity_id": {
                    "type": "string"
                  },
                  "weight": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1
                  },
                  "field": {
                    "type": "string",
                    "default": "role"
                  },
                  "new_value": {
                    "type": "string"
                  },
                  "created_by": {
                    "format": "uuid",
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Override created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "get": {
        "operationId": "listOverrides",
        "summary": "List overrides for a deal",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Overrides list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverridesListResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/deals/{deal_id}/analysis/latest": {
      "get": {
        "operationId": "getLatestAnalysis",
        "summary": "Get latest LIVE_DRAFT analysis run",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest analysis run (null if none)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestAnalysisResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/deals/{deal_id}/export": {
      "post": {
        "operationId": "exportSnapshot",
        "summary": "Run pipeline and create immutable snapshot (idempotent by sha256)",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export result (idempotent)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                },
                "examples": {
                  "successful_export": {
                    "summary": "Successful idempotent export",
                    "value": {
                      "analysis_run": {
                        "id": "a1b2c3d4-0000-0000-0000-000000000001",
                        "deal_id": "d1e2f3a4-0000-0000-0000-000000000001",
                        "state": "LIVE_DRAFT",
                        "schema_version": "v1",
                        "config_version": "v1",
                        "run_trigger": "parse_complete",
                        "non_transfer_abs_total_cents": 150000,
                        "classified_abs_total_cents": 150000,
                        "coverage_pct_bp": 10000,
                        "missing_month_count": 0,
                        "missing_month_penalty_bp": 0,
                        "override_penalty_bp": 0,
                        "reconciliation_status": "NOT_RUN",
                        "reconciliation_pct_bp": null,
                        "base_confidence_bp": 10000,
                        "final_confidence_bp": 10000,
                        "tier": "High",
                        "tier_capped": true,
                        "raw_transaction_hash": "abc123...",
                        "transfer_links_hash": "def456...",
                        "entities_hash": "ghi789...",
                        "overrides_hash": "jkl012..."
                      },
                      "snapshot": {
                        "id": "s1a2b3c4-0000-0000-0000-000000000001",
                        "deal_id": "d1e2f3a4-0000-0000-0000-000000000001",
                        "financial_state_hash": "9d6c0b4c8c6c4f8b2f1e3d2c1a0b9e8d7c6b5a4f3e2d1c0b9a8f7e6d5c4b3a2f",
                        "sha256_hash": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
                        "schema_version": "v1",
                        "config_version": "v1",
                        "created_at": "2024-06-15T12:00:00Z"
                      }
                    }
                  },
                  "denom_zero": {
                    "summary": "Zero denominator (all transfers)",
                    "value": {
                      "analysis_run": {
                        "coverage_pct_bp": 0,
                        "reconciliation_status": "NOT_RUN",
                        "final_confidence_bp": 0,
                        "tier": "Low",
                        "tier_capped": false
                      },
                      "snapshot": {
                        "financial_state_hash": "...",
                        "sha256_hash": "..."
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/deals/{deal_id}/snapshots": {
      "get": {
        "operationId": "listSnapshots",
        "summary": "List snapshots for a deal",
        "parameters": [
          {
            "name": "deal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshots list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotsListResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/snapshots/{snapshot_id}": {
      "get": {
        "operationId": "getSnapshot",
        "summary": "Get a single snapshot by id",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDetailResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DealResponse": {
        "type": "object",
        "required": [
          "deal"
        ],
        "properties": {
          "deal": {
            "$ref": "#/components/schemas/Deal"
          }
        }
      },
      "DealsListResponse": {
        "type": "object",
        "required": [
          "deals"
        ],
        "properties": {
          "deals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Deal"
            }
          }
        }
      },
      "DealDetailResponse": {
        "type": "object",
        "required": [
          "deal",
          "analysis_runs",
          "snapshots"
        ],
        "properties": {
          "deal": {
            "$ref": "#/components/schemas/Deal"
          },
          "analysis_runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisRun"
            }
          },
          "snapshots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Snapshot"
            }
          }
        }
      },
...
      "IngestionResponse": {
        "type": "object",
        "required": [
          "ingestion"
        ],
        "properties": {
          "ingestion": {
            "$ref": "#/components/schemas/IngestionResult"
          }
        }
      },
...
      "DocumentsListResponse": {
        "type": "object",
        "required": [
          "documents"
        ],
        "properties": {
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Document"
            }
          }
        }
      },
...
      "OverrideResponse": {
        "type": "object",
        "required": [
          "override"
        ],
        "properties": {
          "override": {
            "$ref": "#/components/schemas/Override"
          }
        }
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "deal_not_found": {
                "value": {
                  "error_code": "NOT_FOUND",
                  "error_message": "Deal abc-123 not found",
                  "next_action": null,
                  "details": {}
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missing_param": {
                "value": {
                  "error_code": "BAD_REQUEST",
                  "error_message": "created_by query parameter is required",
                  "next_action": null,
                  "details": {}
                }
              }
            }
          }
        }
      },
      "CurrencyMismatch": {
        "description": "Document currency does not match deal currency",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "mismatch": {
                "value": {
                  "error_code": "CURRENCY_MISMATCH",
                  "error_message": "Document currency does not match deal currency",
                  "next_action": "upload_new_file",
                  "details": {}
                }
              }
            }
          }
        }
      },
      "InvalidSchema": {
        "description": "File schema is invalid or unparseable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missing_cols": {
                "value": {
                  "error_code": "INVALID_SCHEMA",
                  "error_message": "Missing required columns: amount, date",
                  "next_action": "fix_format",
                  "details": {}
                }
              }
            }
          }
        }
      }
    }
  }
}